Split tests into separate classes (#1504)
