+++ tests/Microsoft.DotNet.Docker.Tests/CommonRuntimeImageTests.cs
@@ -1,0 +1,46 @@
+﻿// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+using System.Collections.Generic;
+using System.Linq;
+using Xunit;
+using Xunit.Abstractions;
+
+namespace Microsoft.DotNet.Docker.Tests
+{
+    public abstract class CommonRuntimeImageTests : ImageTests
+    {
+        protected CommonRuntimeImageTests(ITestOutputHelper outputHelper)
+            : base(outputHelper)
+        {
+        }
+
+        public static IEnumerable<object[]> GetImageData()
+        {
+            return TestData.GetImageData()
+                .Select(imageData => new object[] { imageData });
+        }
+
+        [LinuxImageTheory]
+        [MemberData(nameof(GetImageData))]
+        public void VerifyInsecureFiles(ImageData imageData)
+        {
+            base.VerifyCommonInsecureFiles(imageData);
+        }
+
+        protected void VerifyCommonEnvironmentVariables(ImageData imageData)
+        {
+            List<EnvironmentVariableInfo> variables = new List<EnvironmentVariableInfo>();
+            variables.AddRange(GetCommonEnvironmentVariables());
+            variables.Add(new EnvironmentVariableInfo("ASPNETCORE_URLS", "http://+:80"));
+
+            if (imageData.OS.StartsWith(OS.AlpinePrefix))
+            {
+                variables.Add(new EnvironmentVariableInfo("DOTNET_SYSTEM_GLOBALIZATION_INVARIANT", "true"));
+            }
+
+            EnvironmentVariableInfo.Validate(variables, ImageType, imageData, DockerHelper);
+        }
+    }
+}
