+++ tests/Microsoft.DotNet.Docker.Tests/OS.cs
@@ -6,18 +6,26 @@ namespace Microsoft.DotNet.Docker.Tests
 {
     public class OS
     {
+        // Alpine
         public const string Alpine39 = "alpine3.9";
         public const string Alpine310 = "alpine3.10";
-        public const string Bionic = "bionic";
+
+        // Debian
         public const string Buster = "buster";
         public const string BusterSlim = "buster-slim";
-        public const string Disco = "disco";
         public const string Stretch = "stretch";
         public const string StretchSlim = "stretch-slim";
+
+        // Ubuntu
+        public const string Bionic = "bionic";
+        public const string Disco = "disco";
+
+        // Windows
         public const string NanoServer1809 = "nanoserver-1809";
         public const string NanoServer1903 = "nanoserver-1903";
         public const string NanoServer1909 = "nanoserver-1909";
 
+        // Helpers
         public const string AlpinePrefix = "alpine";
         public const string SlimSuffix = "-slim";
     }
