+++ tests/Microsoft.DotNet.Docker.Tests/RuntimeDepsImageTests.cs
@@ -1,0 +1,26 @@
+﻿// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+using Xunit;
+using Xunit.Abstractions;
+
+namespace Microsoft.DotNet.Docker.Tests
+{
+    public class RuntimeDepsImageTests : CommonRuntimeImageTests
+    {
+        public RuntimeDepsImageTests(ITestOutputHelper outputHelper)
+            : base(outputHelper)
+        {
+        }
+
+        protected override DotNetImageType ImageType => DotNetImageType.Runtime_Deps;
+
+        [LinuxImageTheory]
+        [MemberData(nameof(GetImageData))]
+        public void VerifyEnvironmentVariables(ImageData imageData)
+        {
+            base.VerifyCommonEnvironmentVariables(imageData);
+        }
+    }
+}
