+++ tests/Microsoft.DotNet.Docker.Tests/TestData.cs
@@ -1,0 +1,92 @@
+﻿// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text.RegularExpressions;
+using static Microsoft.DotNet.Docker.Tests.ImageVersion;
+
+namespace Microsoft.DotNet.Docker.Tests
+{
+    public static class TestData
+    {
+        private static readonly ImageData[] s_linuxTestData =
+        {
+            new ImageData { Version = V2_1, OS = OS.StretchSlim,  Arch = Arch.Amd64 },
+            new ImageData { Version = V2_1, OS = OS.Bionic,       Arch = Arch.Amd64 },
+            new ImageData { Version = V2_1, OS = OS.Alpine39,     Arch = Arch.Amd64 },
+            new ImageData { Version = V2_1, OS = OS.Alpine310,    Arch = Arch.Amd64 },
+            new ImageData { Version = V2_1, OS = OS.StretchSlim,  Arch = Arch.Arm },
+            new ImageData { Version = V2_1, OS = OS.Bionic,       Arch = Arch.Arm },
+            new ImageData { Version = V2_2, OS = OS.StretchSlim,  Arch = Arch.Amd64 },
+            new ImageData { Version = V2_2, OS = OS.Bionic,       Arch = Arch.Amd64 },
+            new ImageData { Version = V2_2, OS = OS.Alpine39,     Arch = Arch.Amd64 },
+            new ImageData { Version = V2_2, OS = OS.Alpine310,    Arch = Arch.Amd64 },
+            new ImageData { Version = V2_2, OS = OS.StretchSlim,  Arch = Arch.Arm },
+            new ImageData { Version = V2_2, OS = OS.Bionic,       Arch = Arch.Arm },
+            new ImageData { Version = V3_0, OS = OS.BusterSlim,   Arch = Arch.Amd64 },
+            new ImageData { Version = V3_0, OS = OS.Disco,        Arch = Arch.Amd64 },
+            new ImageData { Version = V3_0, OS = OS.Bionic,       Arch = Arch.Amd64 },
+            new ImageData { Version = V3_0, OS = OS.Alpine39,     Arch = Arch.Amd64 },
+            new ImageData { Version = V3_0, OS = OS.Alpine310,    Arch = Arch.Amd64 },
+            new ImageData { Version = V3_0, OS = OS.BusterSlim,   Arch = Arch.Arm },
+            new ImageData { Version = V3_0, OS = OS.Disco,        Arch = Arch.Arm },
+            new ImageData { Version = V3_0, OS = OS.Bionic,       Arch = Arch.Arm },
+            new ImageData { Version = V3_0, OS = OS.BusterSlim,   Arch = Arch.Arm64 },
+            new ImageData { Version = V3_0, OS = OS.Disco,        Arch = Arch.Arm64 },
+            new ImageData { Version = V3_0, OS = OS.Bionic,       Arch = Arch.Arm64 },
+            new ImageData { Version = V3_0, OS = OS.Alpine39,     Arch = Arch.Arm64,    SdkOS = OS.Buster },
+            new ImageData { Version = V3_0, OS = OS.Alpine310,    Arch = Arch.Arm64,    SdkOS = OS.Buster },
+            new ImageData { Version = V3_1, OS = OS.BusterSlim,   Arch = Arch.Amd64 },
+            new ImageData { Version = V3_1, OS = OS.Bionic,       Arch = Arch.Amd64 },
+            new ImageData { Version = V3_1, OS = OS.Alpine310,    Arch = Arch.Amd64 },
+            new ImageData { Version = V3_1, OS = OS.BusterSlim,   Arch = Arch.Arm },
+            new ImageData { Version = V3_1, OS = OS.Bionic,       Arch = Arch.Arm },
+            new ImageData { Version = V3_1, OS = OS.BusterSlim,   Arch = Arch.Arm64 },
+            new ImageData { Version = V3_1, OS = OS.Bionic,       Arch = Arch.Arm64 },
+            new ImageData { Version = V3_1, OS = OS.Alpine310,    Arch = Arch.Arm64,    SdkOS = OS.Buster },
+        };
+        private static readonly ImageData[] s_windowsTestData =
+        {
+            new ImageData { Version = V2_1, OS = OS.NanoServer1809, Arch = Arch.Amd64 },
+            new ImageData { Version = V2_1, OS = OS.NanoServer1903, Arch = Arch.Amd64 },
+            new ImageData { Version = V2_1, OS = OS.NanoServer1909, Arch = Arch.Amd64 },
+            new ImageData { Version = V2_2, OS = OS.NanoServer1809, Arch = Arch.Amd64 },
+            new ImageData { Version = V2_2, OS = OS.NanoServer1809, Arch = Arch.Arm },
+            new ImageData { Version = V2_2, OS = OS.NanoServer1903, Arch = Arch.Amd64 },
+            new ImageData { Version = V2_2, OS = OS.NanoServer1909, Arch = Arch.Amd64 },
+            new ImageData { Version = V3_0, OS = OS.NanoServer1809, Arch = Arch.Amd64 },
+            new ImageData { Version = V3_0, OS = OS.NanoServer1809, Arch = Arch.Arm },
+            new ImageData { Version = V3_0, OS = OS.NanoServer1903, Arch = Arch.Amd64 },
+            new ImageData { Version = V3_0, OS = OS.NanoServer1909, Arch = Arch.Amd64 },
+            new ImageData { Version = V3_1, OS = OS.NanoServer1809, Arch = Arch.Amd64 },
+            new ImageData { Version = V3_1, OS = OS.NanoServer1809, Arch = Arch.Arm },
+            new ImageData { Version = V3_1, OS = OS.NanoServer1903, Arch = Arch.Amd64 },
+            new ImageData { Version = V3_1, OS = OS.NanoServer1909, Arch = Arch.Amd64 },
+        };
+
+        public static IEnumerable<ImageData> GetImageData()
+        {
+            string archFilterPattern = GetFilterRegexPattern("IMAGE_ARCH_FILTER");
+            string osFilterPattern = GetFilterRegexPattern("IMAGE_OS_FILTER");
+            string versionFilterPattern = GetFilterRegexPattern("IMAGE_VERSION_FILTER");
+
+            // Filter out test data that does not match the active architecture, os and version filters.
+            return (DockerHelper.IsLinuxContainerModeEnabled ? s_linuxTestData : s_windowsTestData)
+                .Where(imageData => archFilterPattern == null
+                    || Regex.IsMatch(Enum.GetName(typeof(Arch), imageData.Arch), archFilterPattern, RegexOptions.IgnoreCase))
+                .Where(imageData => osFilterPattern == null
+                    || Regex.IsMatch(imageData.OS, osFilterPattern, RegexOptions.IgnoreCase))
+                .Where(imageData => versionFilterPattern == null
+                    || Regex.IsMatch(imageData.VersionString, versionFilterPattern, RegexOptions.IgnoreCase));
+        }
+
+        private static string GetFilterRegexPattern(string filterEnvName)
+        {
+            string filter = Environment.GetEnvironmentVariable(filterEnvName);
+            return filter != null ? $"^{Regex.Escape(filter).Replace(@"\*", ".*").Replace(@"\?", ".")}$" : null;
+        }
+    }
+}
